Don't use spaces to seperate labels
The current behavior of label is to interpret it's value as a space
seperated list of labels like so:

- hasTheWord: enterprise
  label: innovative cloud growth hacking

"label" gets expanded to the labels "innovative", "cloud", "growth", and
"hacking".

This is broken. Mainly because yaml lists would be a way more obvious
choice here, but also because labels can contain spaces.

So: store multiple tags in a normal yaml list.